// webapp/src/types/backstage_rhs.ts
export enum BackstageRHSSection {
    TaskInbox = 'task-inbox',

    RunInfo = 'run-info',
    RunTimeline = 'run-timeline',
    RunStatusUpdates = 'run-status-updates',
    RunParticipants = 'run-participants',
}

export enum BackstageRHSViewMode {
    Overlap = 'overlap',
    Shared = 'shared',
}